feat: #0015 Progress Bar for user profile

+++ src/main/java/pe/memo/memoflashcardsbe/repository/ILessonRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import pe.memo.memoflashcardsbe.repository.entities.Lesson;
 
-public interface ILessonRepository extends JpaRepository<Lesson, Integer>, PagingAndSortingRepository<Lesson, Integer> {
+public interface ILessonRepository extends JpaRepository<Lesson, Long>, PagingAndSortingRepository<Lesson, Long> {
     Page<Lesson> findAllByDeletedAtIsNull(Pageable pageable);
 }
